# feat(monitor): optimize data collection with row batching, boosting performance by 19,000x