--- a/.pylintrc
+++ b/.pylintrc
@@ -9,6 +9,7 @@ disable=
     too-few-public-methods, # enum classes don't normally have methods
     too-many-instance-attributes,
     too-many-arguments,
+    too-many-positional-arguments,
     import-error,
     attribute-defined-outside-init,
     redefined-outer-name