--- a/coleman4hcs/tests/utils/test_monitor.py
+++ b/coleman4hcs/tests/utils/test_monitor.py
@@ -356,6 +356,9 @@ def test_save_with_non_empty_temp_rows(tmp_path, monitor_collector, mock_scenari
 @pytest.mark.performance
 @pytest.mark.parametrize("num_records", [1000, 10_000, 100_000])
 def test_temp_limit_performance(benchmark, num_records):
+    """
+    Performance benchmark for data collection and handling large temporary buffers.
+    """
     mock_metric = MagicMock()
     mock_metric.detected_failures = 5
     mock_metric.undetected_failures = 3