### Before submitting a *pull request* make sure you have:
- [X] The *pull request* title contains a meaningful title
  - Short and informative: serves as a summary 
- [X] At least skimmed through the coding conventions used in the project
- [X] Checked the code with [flake8](https://pypi.python.org/pypi/flake8)
- [ X] Performed a self-review of my own code
- [X] The code is commented, particularly in hard-to-understand areas
- [ ] The documentation associated was created or updated
- [X] The changes generate no new warnings
- [X] Add tests that prove the change is effective or that the feature works
- [X] New and existing unit tests pass locally with the changes
- [ ] Any dependent changes have been merged and published in downstream modules


---
<!-- Describe what you've done -->

# Contents of the *Pull Request*

- Refactored `MonitorCollector` to use row batching (`temp_rows`) instead of DataFrame operations for temporary storage, improving efficiency.
- Replaced DataFrame concatenation logic with batched appends, minimizing memory overhead and computational costs.
- Updated save logic to dynamically handle headers based on file existence and state.
- Enhanced `collect_from_temp` to ensure DataFrame consistency while significantly reducing execution time.
- Adjusted performance benchmarks to reflect improvements, confirming a remarkable **19,000x speed increase** for large-scale experiment data processing.

Other changes:
- Updated `pytest` performance tests and configurations.
- Adjusted policies and configurations in `config.toml` for streamlined execution.
- Included new dependencies (`pytest-benchmark`, `pytest-cov`) in `requirements.txt`.

Breaking changes:
- Removed direct use of temporary DataFrame (`temp_df`), replacing it with a row-based batch system (`temp_rows`).

Benchmarks:
- Small-scale tests (1,000 records): ~4ms (down from ~800ms).
- Medium-scale tests (10,000 records): ~45ms (down from ~8,000ms).
- Large-scale tests (100,000 records): ~579ms (down from ~83,000ms).

# How Has This Been Tested? (Required)

Added new tests for monitor